Dislplay resources dynamically
