+++ cinnamon-platform/cinnamon-frontend/src/app/core/components/navigation/navigation.component.html
@@ -56,7 +56,7 @@
     </div>
 
     <app-project-settings></app-project-settings>
-    <app-project-export></app-project-export>
+    <app-project-export *ngIf="userService.isAuthenticated()"></app-project-export>
 </div>
